Use 0 as down/up throughput defaults
Even though the docs suggest that -1 disables down/upload throughput
throttling, this does not seem to work. Whenever I pass -1 as the
default, the timeline will not generate. 0, on the other hand, seems to
have the same effect as not introducing any throttling. The defaults are
updated to reflect this.