Remove spread operator
This project promises compatibility for Node 8+. Unfortunately, it
appears that Node 8.0.x, 8.1.x, and 8.2.x do not support the spread
operator. Running the `generate` command on those Node version throws an
error. This commit removes the spread operator to honor that promise.

There isn't a test that could have been written to find this problem.
Rather, we need to be testing in different versions of Node. This commit
also adds new versions of Node to test.

Closes #10