Resolve test errors after dependency updates
Jest updated the JSDOM dependency to 11.12+ in a recent update. This
update causes Jest to throw a security error about localStorage when the
tests are assumed to run in the browser and no `testURL` is defined.

Since our tests are not run in the browser, a workaround for this issue
is to explicitly define the environment as `node`. Hopefully this
protects us from weird edge cases in the future.

Refs: facebook/jest#6766, jsdom/jsdom#2304