Update naming convention for object args
Yargs seems to choke on object args using the dot notion. For instance,
the following input:

--foo.bar-biz.baz=3

will result in the following options object:

{
  foo: {
    'bar-biz': {
      baz: 3
    }
  },
  fooBarBizBaz: 3
}

Note how yargs handles camelcase differently in the flattened string
versus the object. I'm worried that this will lead to a lot of
confusion. As a result, I want to remove all dot notion and instead
pass in `-` between strings to denote namespaces.

This commit also fixes up the tests and makes sure the help text is
consistent.